+++ src/simulator.cpp
@@ -283,7 +283,7 @@ int Simulator::eventRoutine(void) {
     this->bitRate = bitRateVariable.getNextIntValue();
     this->rtnAllocation = this->controller->assignConnection(
         this->src, this->dst, this->bitRates[this->bitRate],
-        this->currentEvent.getIdConnection());
+        this->currentEvent.getIdConnection(), this->clock);
     if (this->rtnAllocation == ALLOCATED) {
       nextEventTime = this->clock + this->departVariable.getNextValue();
       for (std::list<Event>::reverse_iterator pos = this->events.rbegin();
@@ -298,7 +298,8 @@ int Simulator::eventRoutine(void) {
       this->allocatedConnections++;
     }
   } else if (this->currentEvent.getType() == DEPARTURE) {
-    this->controller->unassignConnection(this->currentEvent.getIdConnection());
+    (this->controller->*(this->controller->unassignConnection))(
+        this->currentEvent.getIdConnection(), this->clock);
   }
   this->events.pop_front();
   return this->rtnAllocation;
@@ -359,7 +360,7 @@ double Simulator::agrestiCI() {
   double np = this->getAllocatedProbability();
   int n = this->numberOfConnections;
 
-  np = np * ((n * (this->allocatedConnections + (this->zScoreEven/2))) /
+  np = np * ((n * (this->allocatedConnections + (this->zScoreEven / 2))) /
              (this->allocatedConnections * (n + this->zScoreEven)));
 
   double p = 1 - np;
@@ -408,10 +409,16 @@ void Simulator::initZScore(void) {
   this->zScore = actual;
 }
 
+void Simulator::setUnassignCallback(void (*callbackFunction)(Connection, double,
+                                                             Network *)) {
+  this->controller->setUnassignCallback(callbackFunction);
+}
+
 void Simulator::initZScoreEven(void) {
-  double zEven = pow(this->zScore,2);
-  zEven = floorf(zEven*1000)/1000; //Redondeo a la centesima, similar a error
-  zEven = ceil(zEven/2)*2;
+  double zEven = pow(this->zScore, 2);
+  zEven =
+      floorf(zEven * 1000) / 1000;  // Redondeo a la centesima, similar a error
+  zEven = ceil(zEven / 2) * 2;
 
   this->zScoreEven = zEven;
 }